requirements_mlx_lm.txt
qdrant_client
PyYAML
requests
qdrant-client
pydantic
beautifulsoup4
pytest
streamlit
mlx-lm
ollama
-e file:/Users/henrikynsilehto/Desktop/agent_framework
sentence_transformers
